Move getAcceptedLanguagesFromHeader to the server
